+++ MudBlazor.PdfViewer.Docs/Code/OrientationExample.cs
@@ -1,0 +1,9 @@
+namespace MudBlazor.PdfViewer.Docs.Code;
+
+public static class OrientationExample
+{
+    public static string Example =>
+@"<MudPdfViewer 
+  Orientation=""MudBlazorPdf.Orientation.Landscape""
+  Url=""https://raw.githubusercontent.com/tgothorp/MudBlazor.PdfViewer/refs/heads/main/files/test_pdf_document.pdf""/>";
+}