+++ MudBlazor.PdfViewer.Docs/Code/ScrollExample.cs
@@ -1,0 +1,9 @@
+namespace MudBlazor.PdfViewer.Docs.Code;
+
+public static class ScrollExample
+{
+    public static string Scrolling => @"
+<MudPdfViewer
+  SinglePageMode=""false""
+  Url=""https://raw.githubusercontent.com/tgothorp/MudBlazor.PdfViewer/refs/heads/main/files/test_pdf_document.pdf""/>";
+}