+++ MudBlazor.PdfViewer/PdfError.cs
@@ -1,0 +1,7 @@
+namespace MudBlazorPdf;
+
+internal record PdfError
+{
+    public required PdfErrorType ErrorType { get; init; }
+    public required string? Message { get; set; }
+}