+++ MudBlazor.PdfViewer/PdfErrorType.cs
@@ -1,0 +1,7 @@
+namespace MudBlazorPdf;
+
+internal enum PdfErrorType
+{
+    PasswordRequired,
+    Error,
+}