+++ MudBlazor.PdfViewer/PdfInterop.cs
@@ -1,28 +1,20 @@
 using Microsoft.JSInterop;
 
-namespace MudBlazorPdf;
-
 internal class PdfInterop(IJSRuntime jsRuntime) : IAsyncDisposable
 {
-    private readonly Lazy<Task<IJSObjectReference>> js = 
-        new(() => jsRuntime.InvokeAsync<IJSObjectReference>("import", "./_content/Gotho.MudBlazor.PdfViewer/mudpdf.js").AsTask());
+    private readonly Lazy<Task<IJSObjectReference>> js =
+        new(() => jsRuntime.InvokeAsync<IJSObjectReference>("import", "./_content/Gotho.MudBlazor.PdfViewer/mudpdfviewer.js").AsTask());
+
+    public async Task InitializeAsync(object objRef, string elementId, string documentUrl, double scale, double rotation, bool singlePageMode, string? password = null)
+    {
+        var module = await js.Value;
+        await module.InvokeVoidAsync("init", objRef, elementId, documentUrl, scale, rotation, singlePageMode, password);
+    }
 
     public async Task FirstPageAsync(object objRef, string elementId)
     {
         var module = await js.Value;
         await module.InvokeVoidAsync("firstPage", objRef, elementId);
-    }
-
-    public async Task GotoPageAsync(object objRef, string elementId, int gotoPageNum)
-    {
-        var module = await js.Value;
-        await module.InvokeVoidAsync("gotoPage", objRef, elementId, gotoPageNum);
-    }
-
-    public async Task InitializeAsync(object objRef, string elementId, double scale, double rotation, string url, bool showThumbs)
-    {
-        var module = await js.Value;
-        await module.InvokeVoidAsync("initialize", objRef, elementId, scale, rotation, url, showThumbs);
     }
 
     public async Task LastPageAsync(object objRef, string elementId)
@@ -31,22 +23,28 @@ internal class PdfInterop(IJSRuntime jsRuntime) : IAsyncDisposable
         await module.InvokeVoidAsync("lastPage", objRef, elementId);
     }
 
-    public async Task NextPageAsync(object objRef, string elementId)
-    {
-        var module = await js.Value;
-        await module.InvokeVoidAsync("nextPage", objRef, elementId);
-    }
-
     public async Task PreviousPageAsync(object objRef, string elementId)
     {
         var module = await js.Value;
         await module.InvokeVoidAsync("previousPage", objRef, elementId);
     }
 
-    public async Task PrintAsync(object objRef, string elementId, string url)
+    public async Task NextPageAsync(object objRef, string elementId)
     {
         var module = await js.Value;
-        await module.InvokeVoidAsync("print", objRef, elementId, url);
+        await module.InvokeVoidAsync("nextPage", objRef, elementId);
+    }
+
+    public async Task GotoPageAsync(object objRef, string elementId, int gotoPageNum)
+    {
+        var module = await js.Value;
+        await module.InvokeVoidAsync("goToPage", objRef, elementId, gotoPageNum);
+    }
+
+    public async Task ZoomInOutAsync(object objRef, string elementId, double scale)
+    {
+        var module = await js.Value;
+        await module.InvokeVoidAsync("zoom", objRef, elementId, scale);
     }
 
     public async Task RotateAsync(object objRef, string elementId, double rotation)
@@ -55,19 +53,12 @@ internal class PdfInterop(IJSRuntime jsRuntime) : IAsyncDisposable
         await module.InvokeVoidAsync("rotate", objRef, elementId, rotation);
     }
 
-    public async Task ZoomInOutAsync(object objRef, string elementId, double scale)
+    public async Task PrintDocumentAsync(object objRef, string elementId)
     {
         var module = await js.Value;
-        await module.InvokeVoidAsync("zoomInOut", objRef, elementId, scale);
-    }
-    
-    public async Task RenderThumbs(object objRef, string elementId)
-    {
-        var module = await js.Value;
-        await module.InvokeVoidAsync("renderThumbnails", objRef, elementId);
+        await module.InvokeVoidAsync("printDocument", objRef, elementId);
     }
 
-    
     public async ValueTask DisposeAsync()
     {
         if (js.IsValueCreated)